ShareSheet handles mail case.
Now ShareSheet takes only URL as a shareItem parameters and based on the user decision it provides you with entry email text or just a link to session when copying only.
