--- AirCasting/Utils/SharePage.swift
@@ -1,21 +1,0 @@
-// Created by Lunar on 08/07/2021.
-//
-
-import MessageUI
-import SwiftUI
-
-struct ActivityViewController: UIViewControllerRepresentable {
-    typealias Callback = (_ activityType: UIActivity.ActivityType?, _ completed: Bool, _ returnedItems: [Any]?, _ error: Error?) -> Void
-    
-    var itemsToShare: [Any]
-    var servicesToShareItem: [UIActivity]? = nil
-    var completion: Callback? = nil
-    
-    func makeUIViewController(context: UIViewControllerRepresentableContext<ActivityViewController>) -> UIActivityViewController {
-        let controller = UIActivityViewController(activityItems: itemsToShare, applicationActivities: servicesToShareItem)
-        controller.completionWithItemsHandler = completion
-        return controller
-    }
-    
-    func updateUIViewController(_ uiViewController: UIActivityViewController, context: UIViewControllerRepresentableContext<ActivityViewController>) { }
-}
